backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
pymongo==4.8.0
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.9.0
gunicorn==22.0.0
Pillow==12.2.0
redis==5.0.8
